+++ functions/Copy-DbaCmsRegServer.ps1
@@ -116,7 +116,7 @@ function Copy-DbaCmsRegServer {
                     if ($force -eq $false) {
                         if ($Pscmdlet.ShouldProcess($destinstance, "Checking to see if $groupName exists")) {
                             $copyDestinationGroupStatus.Status = "Skipped"
-                            $copyDestinationGroupStatus.Notes = "Already exists"
+                            $copyDestinationGroupStatus.Notes = "Already exists on destination"
                             $copyDestinationGroupStatus | Select-DefaultView -Property DateTime, SourceServer, DestinationServer, Name, Type, Status, Notes -TypeName MigrationObject
                             Write-Message -Level Verbose -Message "Destination group $groupName exists at destination. Use -Force to drop and migrate."
                         }
@@ -181,7 +181,7 @@ function Copy-DbaCmsRegServer {
                     if ($force -eq $false) {
                         if ($Pscmdlet.ShouldProcess($destinstance, "Checking to see if $instanceName in $groupName exists")) {
                             $copyInstanceStatus.Status = "Skipped"
-                            $copyInstanceStatus.Notes = "Already exists"
+                            $copyInstanceStatus.Notes = "Already exists on destination"
                             $copyInstanceStatus | Select-DefaultView -Property DateTime, SourceServer, DestinationServer, Name, Type, Status, Notes -TypeName MigrationObject
 
                             Write-Message -Level Verbose -Message "Instance $instanceName exists in group $groupName at destination. Use -Force to drop and migrate."
@@ -249,7 +249,7 @@ function Copy-DbaCmsRegServer {
                     if ($force -eq $false) {
                         if ($Pscmdlet.ShouldProcess($destinstance, "Checking to see if subgroup $fromSubGroupName exists")) {
                             $copyGroupStatus.Status = "Skipped"
-                            $copyGroupStatus.Notes = "Already exists"
+                            $copyGroupStatus.Notes = "Already exists on destination"
                             $copyGroupStatus | Select-DefaultView -Property DateTime, SourceServer, DestinationServer, Name, Type, Status, Notes -TypeName MigrationObject
 
                             Write-Message -Level Verbose -Message "Subgroup $fromSubGroupName exists at destination. Use -Force to drop and migrate."