+++ functions/Start-DbaPfDataCollectorSet.ps1
@@ -89,7 +89,7 @@ function Start-DbaPfDataCollectorSet {
         }
     }
     process {
-        if (-not (Test-Windows)) { return }
+        
         
         if (-not $InputObject -or ($InputObject -and (Test-Bound -ParameterName ComputerName))) {
             foreach ($computer in $ComputerName) {