+++ changelog.md
@@ -4,6 +4,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.538] - 2018-11-30
+### Added
+* Update-DbaInstance 🎉
+
+### Fixed
+* ComputerName resolution for fqdn in Connect-*Instance
+* Stop-Function -Continue bug in Set-DbaPrivilege
+
+## [0.9.537] - 2018-11-29
+### Added
+* Invoke-DbaDbccFreeCache
+* Get-DbaDbccUserOption
+* Added PolyBase support to Get-DbaService
+
+## [0.9.535] - 2018-11-29
+### Fixed
+- Backup recoveryfork bug
+- Standardized output for Copy command notes
+- New-DbaAgentJobStep issue with server / SubSystemServer
+
+### Added
+* Get-DbaDbccHelp
+* Get-DbaDbccMemoryStatus
+* Get-DbaDbccProcCache
+
+## [0.9.534] - 2018-11-29
+### Fixed
+- Removed mandatory=$false on parameters because it's implied
+
+### Added
+* Get-DbaAgentServer
+* Set-DbaAgentServer
+* Path parameter to audit copies
+
+
 ## [0.9.533] - 2018-11-27
 ### Fixed
 - Removed mandatory=$false on parameters because it's implied